Changed name of package and output class to match EvaWsClient class
