+++ vcf-dumper/vcf-dumper-lib/src/test/java/uk/ac/ebi/eva/vcfdump/evawsclient/EvaWsClientTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.vcfdump.chromosomewsclient;
+package uk.ac.ebi.eva.vcfdump.evawsclient;
 
 import org.junit.Before;
 import org.junit.Rule;
